Working on Google Maps file.
For some reason, the Google Maps module doesn’t work in the VM but
works in my laptop. I am trying to write the functions out and then go
to office hours and fix the problem with the VM? Also made changes to
Django files as I am working through the tutorial.